requirements.txt
pandas
streamlit
python-dotenv
google-generativeai
